refactor : create DateTime for removing duplicated feature
